Move ErrorResponse class and convert it to a Kotlin data class
