+++ components/proxy/src/main/kotlin/com/hotels/styx/ErrorResponse.kt
@@ -13,29 +13,9 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api;
+package com.hotels.styx
 
 /**
  * Represents the body of an HTTP error response in a standard format.
  */
-public class ErrorResponse {
-
-    private String errorMessage;
-
-    /**
-     * Creates an {@link ErrorResponse} with an error message string.
-     * @param errorMessage the error message
-     */
-    public ErrorResponse(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-
-    /**
-     * Return the error message.
-     * @return the error message.
-     */
-    public String errorMessage() {
-        return errorMessage;
-    }
-
-}
+data class ErrorResponse(val errorMessage:String)
